+++ BatForm/Compatibility.cs
@@ -1,0 +1,7 @@
+﻿namespace BatForm
+{
+	internal class CompatibilityUtility
+	{
+		internal static readonly bool IsZoomLevelLoaded = ModEntry.SHelper.ModRegistry.IsLoaded("thespbgamer.ZoomLevel");
+	}
+}